+++ common/src/terrain/structure.rs
@@ -48,9 +48,10 @@ make_case_elim!(
         Sign(content: Content, ori: u8) = 28,
         KeyholeBars(consumes: String) = 29,
         HaniwaKeyhole(consumes: String) = 30,
-        MapleLeaves = 31,
-        CherryLeaves = 32,
-        AutumnLeaves = 33,
+        TerracottaKeyhole(consumes: String) = 31,
+        MapleLeaves = 32,
+        CherryLeaves = 33,
+        AutumnLeaves = 34,
     }
 );