+++ server/src/migrations/V55__steeltonguedrum.sql
@@ -1,0 +1,3 @@
+UPDATE item
+    SET item_definition_id = 'common.items.tool.instruments.steeltonguedrum'
+    WHERE item_definition_id = 'common.items.tool.instruments.steeldrum';