+++ voxygen/anim/src/golem/mod.rs
@@ -141,6 +141,7 @@ impl<'a> From<&'a Body> for SkeletonAttr {
                 (CoralGolem, _) => (-1.0, 3.0),
                 (Gravewarden, _) => (-2.0, 7.0),
                 (AncientEffigy, _) => (-2.0, 8.0),
+                (Mogwai, _) => (-8.0, 2.0),
             },
             jaw: match (body.species, body.body_type) {
                 (StoneGolem, _) => (0.0, 0.0),
@@ -150,6 +151,7 @@ impl<'a> From<&'a Body> for SkeletonAttr {
                 (CoralGolem, _) => (0.0, 0.0),
                 (Gravewarden, _) => (0.0, 0.0),
                 (AncientEffigy, _) => (0.0, 0.0),
+                (Mogwai, _) => (-6.0, -5.0),
             },
             upper_torso: match (body.species, body.body_type) {
                 (StoneGolem, _) => (0.0, 34.5),
@@ -159,6 +161,7 @@ impl<'a> From<&'a Body> for SkeletonAttr {
                 (CoralGolem, _) => (0.0, 25.0),
                 (Gravewarden, _) => (0.0, 26.5),
                 (AncientEffigy, _) => (0.0, 18.0),
+                (Mogwai, _) => (0.0, 18.0),
             },
             lower_torso: match (body.species, body.body_type) {
                 (StoneGolem, _) => (0.0, -10.5),
@@ -168,6 +171,7 @@ impl<'a> From<&'a Body> for SkeletonAttr {
                 (CoralGolem, _) => (0.0, -11.5),
                 (Gravewarden, _) => (0.0, -4.5),
                 (AncientEffigy, _) => (0.0, -4.5),
+                (Mogwai, _) => (0.0, -4.5),
             },
             shoulder: match (body.species, body.body_type) {
                 (StoneGolem, _) => (8.0, -1.5, 4.0),
@@ -177,6 +181,7 @@ impl<'a> From<&'a Body> for SkeletonAttr {
                 (CoralGolem, _) => (11.0, 1.0, 0.0),
                 (Gravewarden, _) => (8.0, 2.0, 3.0),
                 (AncientEffigy, _) => (8.0, 2.0, 3.0),
+                (Mogwai, _) => (8.0, 2.0, 3.0),
             },
             hand: match (body.species, body.body_type) {
                 (StoneGolem, _) => (12.5, -1.0, -7.0),
@@ -186,6 +191,7 @@ impl<'a> From<&'a Body> for SkeletonAttr {
                 (CoralGolem, _) => (2.5, -1.5, -5.0),
                 (Gravewarden, _) => (8.5, -1.0, -7.0),
                 (AncientEffigy, _) => (8.5, -1.0, -7.0),
+                (Mogwai, _) => (8.5, -1.0, -7.0),
             },
             leg: match (body.species, body.body_type) {
                 (StoneGolem, _) => (4.0, 0.0, -3.5),
@@ -195,6 +201,7 @@ impl<'a> From<&'a Body> for SkeletonAttr {
                 (CoralGolem, _) => (2.5, 0.5, -3.0),
                 (Gravewarden, _) => (1.0, 0.5, -6.0),
                 (AncientEffigy, _) => (1.0, 0.5, -6.0),
+                (Mogwai, _) => (1.0, 0.5, -6.0),
             },
             foot: match (body.species, body.body_type) {
                 (StoneGolem, _) => (3.5, 0.5, -9.5),
@@ -204,6 +211,7 @@ impl<'a> From<&'a Body> for SkeletonAttr {
                 (CoralGolem, _) => (2.5, 1.0, -1.5),
                 (Gravewarden, _) => (3.5, -1.0, -8.5),
                 (AncientEffigy, _) => (3.5, -1.0, -8.5),
+                (Mogwai, _) => (3.5, -1.0, -8.5),
             },
             scaler: match (body.species, body.body_type) {
                 (StoneGolem, _) => 1.5,
@@ -213,6 +221,7 @@ impl<'a> From<&'a Body> for SkeletonAttr {
                 (CoralGolem, _) => 1.0,
                 (Gravewarden, _) => 1.5,
                 (AncientEffigy, _) => 1.0,
+                (Mogwai, _) => 1.0,
             },
             tempo: match (body.species, body.body_type) {
                 (StoneGolem, _) => 1.0,
@@ -222,6 +231,7 @@ impl<'a> From<&'a Body> for SkeletonAttr {
                 (CoralGolem, _) => 1.0,
                 (Gravewarden, _) => 1.0,
                 (AncientEffigy, _) => 1.0,
+                (Mogwai, _) => 1.0,
             },
         }
     }